Add clr 8/16/32-bit d/ai/pi/pd/di/ix/aw/al
Note: The MC68000PRM says: In the MC68000 and MC68008 a memory location
is read before it is cleared. We skip this as Musashi doesn't do that
either, and it's not significant for getting the correct result of the
op.

It might have an effect in the real world if the CPU was connected to
some memory mapped device that changes state when being read from, but
since we are in our emulated world, someone would have to connect our
emulated CPU to emulated hardware that depends on this behaviour, which
seems unlikely.