Rename of all Dn/An reg ops for improved uniqueness
This helps when running QC on Dn/An ops, as the _d op-names is no longer a prefix of the _di, and the _a op-names of _ai, _aw and _al, and thus is possible to test separately from the command line.

I.e, 'cargo test -- --ignored cmpa_16_an' now runs only one addressing mode, not four!

Making _D => _DN and _A => _AN incidentally also improves aligment ;)