Implement real illegal op
Now we make a distinction between the real ILLEGAL instruction, and the
default case, as it is useful to be able to handle "unintended use of
possibly unimplemented instruction" differently from actually wanting
this to happen, by using the ILLEGAL instruction.

We now output a message in the default illegal-handler, but not in the
one handling the ILLEGAL instruction.