Discard a particular QC test when exception occurs
This is because Musashi seems to overspend cycles and continue executing
the first instruction in the exception handler indicated by the
exception vector, whereas r68k stops when running out of cycles, without
executing any instruction in the handler.

We detect exceptions by detecting an exception vector being read, and in
that case, we only check that memory accesses match up, but do not
compare beyond which vector was taken.

If the memory accesses do match, we discard the test, as the CPUs didn't
stop executing at the same location, we cannot compare CPU state. If the
memory accesses do not match up, the test fails.