Improve PQC scripts
Modifies pqc to run all ignored ops using GNU parallel, saving a
report of the results in the folder target/debug/qc-results. A complete
run in the current completion state takes around 30 mins on my machine.

Any parameters to pqc are used to filter the tests to run.

./pqc add sub will run all tests where name includes 'add' or 'sub', one
process per test. The difference compared to ./qc add sub is that the
latter runs one test-process for add and another for sub, and aborts
each on the first error due to mutex poisoning, whereas ./pqc runs
them in separate processes, isolating failures.

Adds find-errs that checks the results for errors, and lists them (lists
the names of stderr-files from parallel with non-zero size, which can be
piped to 'xargs less' to show)