Add interrupt handling
Add support for pluggable InterruptController, right now just a simple
autovectoring, autoresetting one (AutoInterruptController). However,
adding the InterruptController type parameter to Core, caused the need
for change in a lot of irrelevant places, like the signature of the
opcode handlers.

So Core needs to be/implement a trait that hides the selection of
InterruptController or Memory implementations. For now we've (since
earlier) hardcoded Memory to LoggingMem<OpsLogger>, and now created a
type alias for Core pointing to ConfiguredCore<AutoInterruptController>,
which makes the choice of InterruptController non-pluggable except by
changing the lib.