Introduce exception callback for logging/override
When an exception (address error, illegal instruction, interrupts etc.)
occurs, the exception_callback will be called with that exception.

If it returns the same (or any) exception, then the normal exception
handling routines of the m68k will be called (on the returned Err), and
execution will continue at the handler pointed to by the appropriate
exception vector.

If it returns Ok(Cycles(...)), the normal exception handling won't be
invoked, and after deducting the cycles returned, execution simply
continues normally with the next instruction - as if the exception
vector was followed, and the exception routine has returned.

This means you can just start execution at an entry point, without even
setting up exception vectors at all, and for example handle TRAPs and
interrupts as you see fit.