Core will enter halted state on double G0 fault
As we currently only emulate Address Error among the G0 exceptions (Bus
Errors are not yet emulated, and external reset is not handled like an
exception), the only way to halt the core is to get an address error in
the address error exception handler. Then it will assume an
unrecoverable system failure and halt.