Rename Core to TestCore, make impls generic
The Core type alias was only used for testing purposes, and "hardcoded"
use of the LoggingMem and the AutoInterruptController.

Also, it made the impl of the TCore trait, and the impl of Core (now
TestCore) itself, specific to those choices. Now the impls of TCore and
ConfiguredCore are generic over memory and interrupt controller, and
only the constructors are specific to TestCore.