Add instruction callback
This function is invoked after reading an opcode, but before it is
executed. It is passed the mutable Core, the PC where the opcode was
read from, and the opcode read.

It can either return the opcode (possibly modified) as an Ok value, or
an Exception as an Err value.

Note that on the passed core, PC has already advanced by two, ready to
read any extension words needed.